studio/requirements.txt
langchain-core
langchain==0.3.18
langchain-community==0.3.17
langchain-gigachat==0.3.4
python-dotenv==1.0.1
langgraph==0.2.74
pandas==2.2.3
openpyxl==3.1.5
faiss-cpu

